+++ flight_sql/flight_sql_statement.h
@@ -28,6 +28,7 @@ private:
   arrow::flight::sql::FlightSqlClient &sql_client_;
   std::shared_ptr<odbcabstraction::ResultSet> current_result_set_;
   std::shared_ptr<arrow::flight::sql::PreparedStatement> prepared_statement_;
+  std::shared_ptr<arrow::flight::FlightInfo> flight_info_;
   const odbcabstraction::MetadataSettings& metadata_settings_;
   long update_count_ = -1;